// src/api/logistic/config-salesorg-customer-group-salesman-group/types.ts
export interface ListSalesORGCustomerGroupCustomerGroup {
  readonly company_id: string
  readonly sales_org_id: string
  readonly channel_id: string
  readonly customer_group_id: string
  readonly salesman_group_id: string
  readonly min_line: number
  readonly min_qty: number
  readonly uom: string
  readonly uom_type: string
  readonly min_amount: number
  readonly status: number
}